fix: address PR#16 review — restructure draw commands, use geom vec, add viewport
Addresses all 11 review comments:

1. BUILD: Split jpov_interface into sub cc_libraries per class
2. Camera: Add viewport rect (x/y/width/height) for 3D rendering region
3. Vec: Reuse geom::Vec2<float>/Vec3<float> via using alias
4. DrawCommand: Redesign with type-specific structs + order queue
5. Params: Use const& for Vec2/Color parameters
6. DrawRect: Remove filled param, always fill (use DrawPolyline for outline)
7. Vertex/PrimitiveTopology: Removed, replaced by per-type structs
8. Label: Keep as debug_label for AI semantic verification
9. Line endpoint: Square cap (immediate mode, no fancy caps)
10. DrawText3D: Participate in depth testing, rendered as 3D textured quad
11. WindowInfo: Change width/height to float for consistency with InputSnapshot

+++ tools/jpov/interface/window_info.h
@@ -6,13 +6,14 @@
 //
 // 设计原则：
 // - 只包含窗口尺寸信息，不含输入事件（输入事件走 InputSnapshot）
-// - 值为当前帧窗口的实际像素尺寸，不会在帧内变化
+// - 值为当前帧窗口的实际像素尺寸（float 类型，与 InputSnapshot 的鼠标坐标类型一致，
+//   允许子像素精度，但实际窗口分辨率应为整数，仍允许非整数值存在但不保证有意义）
 // - 分辨率变化时下一帧自动更新
 
 #ifndef JPOV_WINDOW_INFO_H_
 #define JPOV_WINDOW_INFO_H_
 
-#include <cstdint>
+#include <cmath>
 
 #include <glog/logging.h>
 
@@ -26,13 +27,13 @@ namespace jpov {
 // Pre-condition: width > 0 && height > 0
 // 窗口最小化或尺寸为零时，框架不调用 OneIteration::Step()
 struct WindowInfo {
-    int width;
-    int height;
+    float width;
+    float height;
 
     // 宽高比（width / height），用于投影矩阵计算
     float AspectRatio() const {
-        CHECK_GT(height, 0);
-        return static_cast<float>(width) / static_cast<float>(height);
+        CHECK_GT(height, 0.0f);
+        return width / height;
     }
 };
